@@ -1,44 +1,31 @@
 {
   "name": "awesomehub-angular-app",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "author": "Mohamed Kholy <contact@mdkholy.me>",
-  "description": "Angular 2 application for AwesomeHub",
-  "main": "src/app/main.ts",
+  "description": "Angular application for AwesomeHub",
+  "main": "src/main.ts",
   "homepage": "https://github.com/awesomehub/angular-app",
   "repository": {
     "type": "git",
     "url": "https://github.com/awesomehub/angular-app.git"
   },
   "engines": {
-    "node": ">=5.0",
-    "yarn": "^0.16.0"
+    "node": ">=8.0",
+    "yarn": "^1.0"
   },
   "scripts": {
-    "start": "webpack-dashboard -- webpack-dev-server --inline --port 8080",
-    "start:raw": "webpack-dev-server --progress --profile --port 8080",
-    "pretest": "yarn run clean-coverage",
-    "test": "karma start",
-    "tslint": "tslint",
-    "lint": "yarn run tslint \"src/**/*.ts\"",
-    "serve": "lite-server -c config/browser-sync/bs-config.js",
-    "prebuild": "yarn run clean-build",
-    "build": "webpack --config config/webpack/webpack.prod.js --progress --profile --bail",
-    "clean": "yarn run clean-build && yarn run clean-coverage",
-    "clean-build": "rimraf dist",
-    "clean-coverage": "rimraf coverage",
-    "pree2e": "yarn run webdriver:update -- --standalone",
-    "e2e": "protractor",
-    "webdriver-manager": "webdriver-manager",
-    "webdriver:start": "yarn run webdriver-manager start",
-    "webdriver:update": "yarn run webdriver-manager update",
-    "coveralls": "coveralls < ./coverage/lcov.info",
-    "stat": "node ./config/dashboard"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "build:production": "ng build --configuration production",
+    "serve": "serve dist",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
   "keywords": [
     "app",
     "angular",
-    "angular2",
-    "webpack",
     "github"
   ],
   "license": "MIT",
@@ -46,80 +33,32 @@
     "url": "https://github.com/awesomehub/angular-app/issues"
   },
   "dependencies": {
-    "@angular/common": "~2.2.3",
-    "@angular/compiler": "~2.2.3",
-    "@angular/compiler-cli": "~2.2.3",
-    "@angular/core": "~2.2.3",
-    "@angular/forms": "~2.2.3",
-    "@angular/http": "~2.2.3",
-    "@angular/platform-browser": "~2.2.3",
-    "@angular/platform-browser-dynamic": "~2.2.3",
-    "@angular/router": "~3.2.3",
-    "@ngtools/webpack": "1.1.6",
-    "@types/core-js": "0.9.34",
-    "@types/jasmine": "2.5.38",
-    "@types/lodash": "4.14.39",
-    "@types/node": "6.0.51",
-    "angular2-router-loader": "^0.3.4",
-    "angular2-template-loader": "^0.6.0",
-    "awesome-typescript-loader": "^2.2.4",
-    "chalk": "^1.1.3",
-    "codelyzer": "1.0.0-beta.4",
-    "compression-webpack-plugin": "0.3.2",
-    "core-js": "^2.4.1",
-    "css-loader": "^0.26.0",
-    "dotenv": "^2.0.0",
-    "extract-text-webpack-plugin": "^2.0.0-beta.4",
-    "file-loader": "^0.9.0",
-    "font-awesome": "^4.7.0",
-    "html-loader": "^0.4.4",
-    "html-webpack-plugin": "^2.24.1",
-    "istanbul-instrumenter-loader": "0.2.0",
-    "jasmine-core": "~2.5.2",
-    "karma": "~1.3.0",
-    "karma-coverage": "~1.1.1",
-    "karma-jasmine": "~1.0.2",
-    "karma-phantomjs-launcher": "~1.0.2",
-    "karma-remap-coverage": "^0.1.2",
-    "karma-webpack": "1.8.0",
-    "lodash": "^4.17.2",
-    "null-loader": "^0.1.1",
-    "phantomjs-prebuilt": "^2.1.13",
-    "postcss-cssnext": "^2.8.0",
-    "postcss-loader": "^1.1.1",
-    "precss": "^1.4.0",
-    "raw-loader": "^0.5.1",
-    "rimraf": "^2.5.4",
-    "rxjs": "5.0.0-beta.12",
-    "style-loader": "^0.13.1",
-    "tslint": "^3.15.1",
-    "tslint-loader": "^2.1.5",
-    "typescript": "2.0.10",
-    "webpack": "2.1.0-beta.27",
-    "webpack-merge": "^0.18.0",
-    "zone.js": "^0.6.26",
-    "copy-webpack-plugin": "^3.0.1",
-    "@ngrx/core": "~1.2.0",
-    "@ngrx/effects": "~2.0.0",
-    "@ngrx/store": "~2.2.1",
-    "@ngrx/router-store": "~1.2.5",
-    "@ngrx/store-devtools": "^3.2.2",
-    "@ngrx/store-log-monitor": "^3.0.2",
-    "ngrx-store-logger": "^0.1.7",
-    "ngrx-store-freeze": "^0.1.0",
-    "material-design-lite": "^1.2.0",
-    "moment": "^2.15.2"
+    "@angular/common": "~8.0.3",
+    "@angular/compiler": "~8.0.3",
+    "@angular/core": "~8.0.3",
+    "@angular/forms": "~8.0.3",
+    "@angular/platform-browser": "~8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.3",
+    "@angular/router": "~8.0.3",
+    "@ngrx/effects": "~8.0.1",
+    "@ngrx/router-store": "~8.0.1",
+    "@ngrx/store": "~8.0.1",
+    "@ngrx/store-devtools": "~8.0.1",
+    "core-js": "^3.12.1",
+    "material-design-lite": "^1.3.0",
+    "moment": "^2.29.1",
+    "rxjs": "~6.5.5",
+    "tslib": "^1.14.1",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angularclass/hmr": "^1.2.2",
-    "@angularclass/hmr-loader": "^3.0.2",
-    "@types/protractor": "1.5.20",
-    "@types/selenium-webdriver": "2.44.28",
-    "compression": "^1.6.2",
-    "lite-server": "^2.2.2",
-    "protractor": "4.0.11",
-    "ts-node": "^1.7.0",
-    "webpack-dashboard": "^0.2.0",
-    "webpack-dev-server": "^2.1.0-beta.11"
+    "@angular-devkit/build-angular": "~0.800.6",
+    "@angular/cli": "~8.0.6",
+    "@angular/compiler-cli": "~8.0.3",
+    "@angular/language-service": "~8.0.3",
+    "@types/node": "~10.12.30",
+    "ts-node": "~7.0.1",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.5"
   }
 }
