@@ -1,12 +1,10 @@
-import 'rxjs/add/operator/withLatestFrom';
-
-import { Component, OnInit, ViewEncapsulation, ChangeDetectionStrategy, ChangeDetectorRef, ViewChild, ElementRef } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Store } from '@ngrx/store';
-import { go, replace, back } from '@ngrx/router-store';
-
+import { Component, OnInit, AfterViewInit, ViewEncapsulation, ChangeDetectionStrategy, ChangeDetectorRef, ViewChild, ElementRef } from '@angular/core';
+import { Location } from '@angular/common';
+import { ActivatedRoute, Router } from '@angular/router';
+import { Store, select } from '@ngrx/store';
+import { withLatestFrom, distinctUntilChanged } from 'rxjs/operators';
 import { AppState } from '../../../app.state';
-import { HeaderBarRouteComponent, getRouterPath } from '../../../core';
+import { HeaderBarRouteComponent, selectRouterState } from '../../../core';
 
 @Component({
   template: `
@@ -21,7 +19,7 @@ import { HeaderBarRouteComponent, getRouterPath } from '../../../core';
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class SearchBarRouteComponent extends HeaderBarRouteComponent implements OnInit {
+export class SearchBarRouteComponent extends HeaderBarRouteComponent implements OnInit, AfterViewInit {
 
   public placeholder: string = 'Search...';
   public query: string = '';
@@ -30,19 +28,18 @@ export class SearchBarRouteComponent extends HeaderBarRouteComponent implements 
   private cancelRoute: string;
   private hasBack: boolean = false;
 
-  @ViewChild('input') private input: ElementRef;
+  @ViewChild('input', { static: false }) private input: ElementRef;
 
   constructor(
+    private router: Router,
     private route: ActivatedRoute,
     private store$: Store<AppState>,
+    private location: Location,
     private cd: ChangeDetectorRef) {
     super();
   }
 
   ngOnInit() {
-    // Prevents input focus getting lost during typing due to component re-initiation
-    this.input.nativeElement.focus();
-
     this.route.data.forEach(({placeholder, searchRoute, cancelRoute, list}) => {
       this.placeholder = placeholder;
       if (list) {
@@ -55,17 +52,24 @@ export class SearchBarRouteComponent extends HeaderBarRouteComponent implements 
       this.cd.markForCheck();
     });
 
-    this.store$
-      .let(getRouterPath())
-      .withLatestFrom(this.route.data, this.route.queryParams)
-      .forEach(([url, {searchRouteMatch}, {q}]) => {
-        const isSearch = new RegExp(searchRouteMatch).test(<string>url);
-        if(!isSearch){
-          this.hasBack = true;
-        }
-        this.query = isSearch ? q || '' : '';
-        this.cd.markForCheck();
-      });
+    this.store$.pipe(
+      select(selectRouterState),
+      distinctUntilChanged(),
+      withLatestFrom(this.route.data)
+    )
+    .forEach(([{ url, queryParams: { q }}, {searchRouteMatch}]) => {
+      const isSearch = new RegExp(searchRouteMatch).test(<string>url);
+      if(!isSearch){
+        this.hasBack = true;
+      }
+      this.query = isSearch ? q || '' : '';
+      this.cd.markForCheck();
+    });
+  }
+
+  ngAfterViewInit() {
+    // Prevents input focus getting lost during typing due to component re-initiation
+    this.input.nativeElement.focus();
   }
 
   search(query: string): void {
@@ -74,23 +78,25 @@ export class SearchBarRouteComponent extends HeaderBarRouteComponent implements 
       return;
     }
 
-    const path = [this.searchRoute];
-    const params = {q: query};
-    const action = this.query === ''
-      ? go(path, params)
-      : replace(path, params);
+    const path = this.searchRoute;
+    const queryParams = { q: query };
+
+    this.query === ''
+      ? this.router.navigate([path], { queryParams })
+      : this.router.navigate([path], { queryParams, replaceUrl: true });
+
     this.query = query;
-    this.store$.dispatch(action);
     this.cd.markForCheck();
   }
 
   cancel(): void {
     this.query = '';
+
     // If search is the very first page then go to the $cancelRoute
-    const action = this.hasBack
-      ? back()
-      : go([this.cancelRoute]);
-    this.store$.dispatch(action);
+    this.hasBack
+      ? this.location.back()
+      : this.router.navigate([this.cancelRoute]);
+
     this.cd.markForCheck();
   }
 }
