@@ -1,3 +1,5 @@
+import { environment } from '../environments/environment';
+
 /**
  * App Configurations
  */
@@ -6,7 +8,7 @@ export class AppConfig {
   static NAME           = 'AwesomeHub';
 
   /** API **/
-  static API_BASE_URL   = ENV.API_URL;
+  static API_BASE_URL   = environment.api;
   static API_LISTS_URL  = `${AppConfig.API_BASE_URL}/lists`;
   static API_LIST_URL   = `${AppConfig.API_BASE_URL}/list`;
 }
