@@ -1,2 +1,1 @@
-export * from './selector';
-export * from './reducers';
+export * from './action';
