@@ -1,2 +1,3 @@
-export * from './reducers';
+export * from './models';
 export * from './selectors';
+export * from './reducers';
