@@ -1,4 +1,4 @@
-import { Directive, ElementRef, Input, Renderer } from '@angular/core';
+import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
 
 import { ListRepoScoreService } from '../../services';
 
@@ -7,14 +7,17 @@ import { ListRepoScoreService } from '../../services';
 })
 export class ListRepoScoreStyleDirective {
 
-  constructor(private repoScoreService: ListRepoScoreService,
-              private el: ElementRef, private renderer: Renderer) { }
+  constructor(
+    private repoScoreService: ListRepoScoreService,
+    private el: ElementRef,
+    private renderer: Renderer2,
+  ) { }
 
   @Input('scoreType') scoreType: string;
   @Input()
   set listRepoScoreStyle(score: number) {
     let colors = this.repoScoreService.getScoreColor(this.scoreType, score);
-    this.renderer.setElementStyle(this.el.nativeElement, 'color', colors.text);
-    this.renderer.setElementStyle(this.el.nativeElement, 'backgroundColor', colors.bg);
+    this.renderer.setStyle(this.el.nativeElement, 'color', colors.text)
+    this.renderer.setStyle(this.el.nativeElement, 'backgroundColor', colors.bg)
   }
 }
