@@ -1,11 +1,6 @@
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/operator/startWith';
-import 'rxjs/add/operator/map';
-import { Observable } from 'rxjs/Observable';
-
 import { Injectable } from '@angular/core';
-
+import { Observable, fromEvent } from 'rxjs';
+import { map, debounceTime, startWith } from 'rxjs/operators';
 import { ScrollSpyData } from './scroll-spy-data.interface';
 
 @Injectable()
@@ -14,17 +9,18 @@ export class ScrollSpyService {
   public scroll$: Observable<Event>;
 
   constructor() {
-    this.scroll$ = Observable.fromEvent<Event>(window, 'scroll');
+    this.scroll$ = fromEvent<Event>(window, 'scroll');
   }
 
   public getScrollData(debounce: number = 100): Observable<ScrollSpyData> {
-    return this.scroll$
+    return this.scroll$.pipe(
       // Emit an initial data so that observers can update their view on pageload
-      .startWith(<Event><any>{target: document})
-      .debounceTime(debounce)
-      .map((event: Event) => {
+      startWith(<Event><any>{target: document}),
+      debounceTime(debounce),
+      map((event: Event) => {
         return this.createScrollData(event.target as HTMLDocument);
-      });
+      })
+    );
   }
 
   public getScrollDataSnapshot(): ScrollSpyData {
